Remove custom parser/use API for parsing in memslap

+++ tests/mem_functions.c
@@ -2892,7 +2892,7 @@ static test_return_t auto_eject_hosts(memcached_st *trash)
   instance->next_retry = time(NULL) + 15;
   memc->next_distribution_rebuild= time(NULL) - 1;
 
-  for (int x= 0; x < 99; x++)
+  for (size_t x= 0; x < 99; x++)
   {
     uint32_t server_idx = memcached_generate_hash(memc, ketama_test_cases[x].key, strlen(ketama_test_cases[x].key));
     test_truth(server_idx != 2);
@@ -2903,7 +2903,7 @@ static test_return_t auto_eject_hosts(memcached_st *trash)
   memc->next_distribution_rebuild= time(NULL) - 1;
   memcached_behavior_set(memc, MEMCACHED_BEHAVIOR_DISTRIBUTION,
                          memc->distribution);
-  for (int x= 0; x < 99; x++)
+  for (size_t x= 0; x < 99; x++)
   {
     uint32_t server_idx = memcached_generate_hash(memc, ketama_test_cases[x].key, strlen(ketama_test_cases[x].key));
     // We re-use instance from above.
